+++ src/nvhiz_vk.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2025, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2022-2026, NVIDIA CORPORATION.  All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * SPDX-FileCopyrightText: Copyright (c) 2022-2025 NVIDIA CORPORATION
+ * SPDX-FileCopyrightText: Copyright (c) 2022-2026 NVIDIA CORPORATION
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -34,9 +34,10 @@ void NVHizVK::TextureInfo::getShaderFactors(float factors[4]) const
   factors[3] = float(usedHeight - 2) / float(height);
 }
 
-float NVHizVK::TextureInfo::getSizeMax() const
+void NVHizVK::TextureInfo::getSize(float size[2]) const
 {
-  return float(std::max(width, height));
+  size[0] = float(width);
+  size[1] = float(height);
 }
 
 void NVHizVK::deinit()